Simplify paths in tox.ini
Tox by default changes current working directory to {toxinidir}.
Therefore, it is not needed in our tox.ini to explicitly state
that element in commands. Even if tox is called from different
directory with -c option, it is safe to rely on relative paths.
So we can omit that part as long as we do not change `changedir`
value in tox.ini directly.